chore: Clean up project structure and add Archlinux PKGBUILDS
