+++ .gitignore
@@ -1,3 +1,4 @@
 dev
 tests
 __pycache__
+**/__pycache__
