+++ src/omen_fan_control/data/driver/dkms.conf
