+++ src/omen_fan_control/data/driver/hooks/90-hp-wmi-omen.hook
