+++ src/omen_fan_control/data/driver/hooks/99-hp-wmi-omen.install
