+++ src/omen_fan_control/data/driver/hooks/zz-hp-wmi-omen
