+++ src/omen_fan_control/data/driver/hp-wmi.c
