+++ src/omen_fan_control/data/driver/install_driver.sh
